feat: - move run_examples to example_study dir - leave run_scenarios script for running any config
