+++ studies/example_study/run_examples.py
@@ -2,133 +2,18 @@
 # external imports
 # ----------------
 import sys
-import traceback
-import pandas as pd
+import os
 import multiprocessing as mp
 
-# src imports
-# -----------
-import src.misc.config as config
-from src.misc.init_modules import load_simulation_environment
-from src.misc.globals import *
+fleetpy_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(fleetpy_path)
 
-
-# main functions
-# --------------
-def run_single_simulation(scenario_parameters):
-    SF = load_simulation_environment(scenario_parameters)
-    if scenario_parameters.get("bugfix", False):
-        try:
-            SF.run()
-        except:
-            traceback.print_exc()
-    else:
-        SF.run()
-
-
-def run_scenarios(constant_config_file, scenario_file, n_parallel_sim=1, n_cpu_per_sim=1, evaluate=1, log_level="info",
-                  keep_old=False, continue_next_after_error=False):
-    """
-    This function combines constant study parameters and scenario parameters.
-    Then it sets up a pool of workers and starts a simulation for each scenario.
-    The required parameters are stated in the documentation.
-
-    :param constant_config_file: this file contains all input parameters that remain constant for a study
-    :type constant_config_file: str
-    :param scenario_file: this file contain all input parameters that are varied for a study
-    :type scenario_file: str
-    :param n_parallel_sim: number of parallel simulation processes
-    :type n_parallel_sim: int
-    :param n_cpu_per_sim: number of cpus for a single simulation
-    :type n_cpu_per_sim: int
-    :param evaluate: 0: no automatic evaluation / != 0 automatic simulation after each simulation
-    :type evaluate: int
-    :param log_level: hierarchical output to the logging file. Possible inputs with hierarchy from low to high:
-            - "verbose": lowest level -> logs everything; even code which could scale exponentially
-            - "debug": standard debugging logger. code which scales exponentially should not be logged here
-            - "info": basic information during simulations (default)
-            - "warning": only logs warnings
-    :type log_level: str
-    :param keep_old: does not start new simulation if result files are already available in scenario output directory
-    :type keep_old: bool
-    :param continue_next_after_error: continue with next simulation if one the simulations threw an error (only SP)
-    :type continue_next_after_error: bool
-    """
-    assert type(n_parallel_sim) == int, "n_parallel_sim must be of type int"
-    # read constant and scenario config files
-    constant_cfg = config.ConstantConfig(constant_config_file)
-    scenario_cfgs = config.ScenarioConfig(scenario_file)
-
-    # set constant parameters from function arguments
-    # TODO # get study name and check if its a studyname
-    const_abs = os.path.abspath(constant_config_file)
-    study_name = os.path.basename(os.path.dirname(os.path.dirname(const_abs)))
-
-    if study_name == "scenarios":
-        print("ERROR! The path of the config files is not longer up to date!")
-        print("See documentation/Data_Directory_Structure.md for the updated directory structure needed as input!")
-        exit()
-    if constant_cfg.get(G_STUDY_NAME) is not None and study_name != constant_cfg.get(G_STUDY_NAME):
-        print("ERROR! {} from constant config is not consistent with study directory: {}".format(constant_cfg[G_STUDY_NAME], study_name))
-        print("{} is now given directly by the folder name !".format(G_STUDY_NAME))
-        exit()
-    constant_cfg[G_STUDY_NAME] = study_name
-
-    constant_cfg["n_cpu_per_sim"] = n_cpu_per_sim
-    constant_cfg["evaluate"] = evaluate
-    constant_cfg["log_level"] = log_level
-    constant_cfg["keep_old"] = keep_old
-
-    # combine constant and scenario parameters into verbose scenario parameters
-    for i, scenario_cfg in enumerate(scenario_cfgs):
-        scenario_cfgs[i] = constant_cfg + scenario_cfg
-
-    # perform simulation(s)
-    print(f"Simulation of {len(scenario_cfgs)} scenarios on {n_parallel_sim} processes with {n_cpu_per_sim} cpus per simulation ...")
-    if n_parallel_sim == 1:
-        for scenario_cfg in scenario_cfgs:
-            if continue_next_after_error:
-                try:
-                    run_single_simulation(scenario_cfg)
-                except:
-                    traceback.print_exc()
-            else:
-                run_single_simulation(scenario_cfg)
-    else:
-        if n_cpu_per_sim == 1:
-            mp_pool = mp.Pool(n_parallel_sim)
-            mp_pool.map(run_single_simulation, scenario_cfgs)
-        else:
-            n_scenarios = len(scenario_cfgs)
-            rest_scenarios = n_scenarios
-            current_scenario = 0
-            while rest_scenarios != 0:
-                if rest_scenarios >= n_parallel_sim:
-                    par_processes = [None for i in range(n_parallel_sim)]
-                    for i in range(n_parallel_sim):
-                        par_processes[i] = mp.Process(target=run_single_simulation, args=(scenario_cfgs[current_scenario],))
-                        current_scenario += 1
-                        par_processes[i].start()
-                    for i in range(n_parallel_sim):
-                        par_processes[i].join()
-                        rest_scenarios -= 1
-                else:
-                    par_processes = [None for i in range(rest_scenarios)]
-                    for i in range(rest_scenarios):
-                        par_processes[i] = mp.Process(target=run_single_simulation, args=(scenario_cfgs[current_scenario],))
-                        current_scenario += 1
-                        par_processes[i].start()
-                    for i in range(rest_scenarios):
-                        par_processes[i].join()
-                        rest_scenarios -= 1
-
-# -------------------------------------------------------------------------------------------------------------------- #
-# ----> you can replace the following part by your respective if __name__ == '__main__' part for run_private*.py <---- #
-# -------------------------------------------------------------------------------------------------------------------- #
-
+from run_scenarios import run_scenarios
 
 # -------------------------------------------------------------------------------------------------------------------- #
 if __name__ == "__main__":
+    """ this script runs all scenarios of the example study. You can use this script as template for your own run_private*.py script to run the scenarios of your own study. 
+    The required parameters are stated in the documentation of the run_scenarios function."""
     mp.freeze_support()
 
     if len(sys.argv) > 1:
