+++ studies/manhattan_case_study/run_manhattan_case_study_example.py
@@ -7,7 +7,7 @@ import traceback
 fleetpy_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.append(fleetpy_path)
 
-from studies.example_study.run_examples import run_scenarios
+from run_scenarios import run_scenarios
 import src.misc.config as config
 from src.misc.globals import *
